move some exceptions to starknet module
