+++ protostar/commands/migrate/migrate_command.py
@@ -10,8 +10,8 @@ from protostar.cli.network_command_util import NetworkCommandUtil
 from protostar.cli.signable_command_util import SignableCommandUtil
 from protostar.migrator import Migrator, MigratorExecutionEnvironment
 from protostar.protostar_exception import ProtostarException
+from protostar.starknet import CheatcodeException
 from protostar.starknet_gateway.gateway_facade_factory import GatewayFacadeFactory
-from protostar.testing.test_environment_exceptions import CheatcodeException
 from protostar.utils.input_requester import InputRequester
 from protostar.utils.log_color_provider import LogColorProvider
