+++ protostar/commands/migrate/migrate_command_test.py
@@ -10,8 +10,8 @@ from starknet_py.net.models import StarknetChainId
 from protostar.commands.migrate.migrate_command import MigrateCommand
 from protostar.migrator import Migrator
 from protostar.protostar_exception import ProtostarException
+from protostar.starknet import CheatcodeException
 from protostar.starknet_gateway import GatewayFacadeFactory
-from protostar.testing.test_environment_exceptions import CheatcodeException
 from protostar.utils.input_requester import InputRequester
