+++ protostar/commands/test/test_result_formatter.py
@@ -2,6 +2,7 @@ from pathlib import Path
 from typing import Callable, Dict, List
 
 from protostar.protostar_exception import UNEXPECTED_PROTOSTAR_ERROR_MSG
+from protostar.starknet import ExceptionMetadata
 from protostar.testing import (
     BrokenFuzzTestCaseResult,
     BrokenTestCaseResult,
@@ -16,7 +17,6 @@ from protostar.testing import (
     UnexpectedBrokenTestSuiteResult,
     format_output_name,
 )
-from protostar.testing.test_environment_exceptions import ExceptionMetadata
 from protostar.utils.log_color_provider import log_color_provider
 
 LogCallback = Callable[[str], None]
