+++ protostar/migrator/cheatcodes/migrator_declare_cheatcode.py
@@ -6,13 +6,13 @@ from typing import Any, Optional
 from starknet_py.net.signer import BaseSigner
 from typing_extensions import Protocol
 
-from protostar.starknet.cheatcode import Cheatcode
-from protostar.starknet_gateway import GatewayFacade
-from protostar.starknet_gateway.gateway_facade import CompilationOutputNotFoundException
-from protostar.testing.test_environment_exceptions import (
+from protostar.starknet import (
+    Cheatcode,
     CheatcodeException,
     KeywordOnlyArgumentCheatcodeException,
 )
+from protostar.starknet_gateway import GatewayFacade
+from protostar.starknet_gateway.gateway_facade import CompilationOutputNotFoundException
 
 from .network_config import CheatcodeNetworkConfig, ValidatedCheatcodeNetworkConfig
