+++ protostar/testing/cheatcodes/mock_call_cheatcode.py
@@ -3,9 +3,7 @@ from typing import Any, Callable, List, Optional
 
 from starkware.starknet.public.abi import AbiType, get_selector_from_name
 
-from protostar.starknet.cheatcode import Cheatcode
-from protostar.starknet.types import AddressType
-from protostar.testing.test_environment_exceptions import CheatcodeException
+from protostar.starknet import AddressType, Cheatcode, CheatcodeException
 from protostar.utils.data_transformer import (
     CairoOrPythonData,
     PythonData,
