+++ protostar/testing/cheatcodes/prepare_cheatcode.py
@@ -6,8 +6,7 @@ from starkware.starknet.core.os.contract_address.contract_address import (
     calculate_contract_address_from_hash,
 )
 
-from protostar.starknet.cheatcode import Cheatcode
-from protostar.testing.test_environment_exceptions import CheatcodeException
+from protostar.starknet import Cheatcode, CheatcodeException
 from protostar.utils.data_transformer import (
     CairoOrPythonData,
     PythonData,
