+++ protostar/testing/cheatcodes/store_cheatcode.py
@@ -2,8 +2,7 @@ from typing import Any, Callable, List, Optional
 
 from starkware.starknet.business_logic.state.state import ContractStorageState
 
-from protostar.starknet.cheatcode import Cheatcode
-from protostar.starknet.storage_var import calc_address
+from protostar.starknet import Cheatcode, calc_address
 
 ADDR_BOUND = 2**251 - 256
