+++ protostar/testing/environments/fuzz_test_execution_environment.py
@@ -10,6 +10,7 @@ from hypothesis.reporting import with_reporter
 from hypothesis.strategies import SearchStrategy
 from starkware.starknet.business_logic.execution.objects import CallInfo
 
+from protostar.starknet import CheatcodeException, ReportedException
 from protostar.starknet.cheatcode import Cheatcode
 from protostar.testing.cheatcodes import AssumeCheatcode, RejectCheatcode
 from protostar.testing.fuzzing.exceptions import HypothesisRejectException
@@ -27,10 +28,6 @@ from protostar.testing.starkware.execution_resources_summary import (
     ExecutionResourcesSummary,
 )
 from protostar.testing.starkware.test_execution_state import TestExecutionState
-from protostar.testing.test_environment_exceptions import (
-    CheatcodeException,
-    ReportedException,
-)
 from protostar.utils.abi import get_function_parameters
 
 from .test_execution_environment import (
