+++ protostar/testing/test_case_runners/test_case_runner.py
@@ -2,14 +2,11 @@ from abc import abstractmethod
 from dataclasses import dataclass
 from typing import Generic, TypeVar
 
+from protostar.starknet import BreakingReportedException, ReportedException
 from protostar.testing.environments.test_execution_environment import (
     TestExecutionResult,
 )
 from protostar.testing.stopwatch import Stopwatch
-from protostar.testing.test_environment_exceptions import (
-    BreakingReportedException,
-    ReportedException,
-)
 from protostar.testing.test_output_recorder import OutputRecorder
 from protostar.testing.test_results import (
     BrokenTestCaseResult,
