+++ protostar/testing/test_context.py
@@ -1,9 +1,7 @@
 from types import SimpleNamespace
 from typing import Any
 
-from protostar.starknet.hint_local import HintLocal
-
-from .test_environment_exceptions import SimpleReportedException
+from protostar.starknet import HintLocal, SimpleReportedException
 
 
 class TestContext(SimpleNamespace):
