+++ tests/integration/migrator/migrator_call_cheatcode_test.py
@@ -6,8 +6,8 @@ from typing import Dict
 import pytest
 
 from protostar.protostar_exception import ProtostarException
+from protostar.starknet import ReportedException
 from protostar.starknet_gateway.starknet_request import StarknetRequest
-from protostar.testing.test_environment_exceptions import ReportedException
 from tests.data.contracts import IDENTITY_CONTRACT
 from tests.integration.conftest import CreateProtostarProjectFixture
 from tests.integration.migrator.conftest import MigrateFixture
